ignore build directory
I was trying to add freeglut to project and used build directory to
create visual studio project using CMake. I needed to do that to
generate .lib and .dll for freeglut to link it against my own program.
Build directory was not already ignored so I added it here. This commit
does not include the freeglut dependency because I need to reinforce
installation steps in permanent memory and reinstall the library from
scratch to document installation steps for the future reference.