modify each vertex
I've read that OpenGL pipeline applies vertex shader for each vertex
that we add to the OpenGL. If that is true, then I should be able to
modify my current vertex values and move x to the right and y to the
bottom. After adding code to do such thing I finally see it for myself.